requirements.txt
# 股票交易策略助手 依赖包
# Stock Trading Strategy Assistant Dependencies

# 数据获取
yfinance>=0.2.36
akshare>=1.14.0
baostock>=0.8.8
requests>=2.31.0

# 数据处理
pandas>=2.1.0
numpy>=1.26.0

# 技术指标
pandas-ta>=0.3.14b

# 可视化
plotly>=5.18.0
matplotlib>=3.8.0

# AI / LLM
anthropic>=0.40.0

# Web UI
streamlit>=1.35.0

# 配置文件解析
pyyaml>=6.0.1

# 其他工具
python-dotenv>=1.0.0
tabulate>=0.9.0
colorama>=0.4.6
rich>=13.7.0
